updated the project structure
