+++ content-searching/src/main/java/org/altumtek/EchoClient.java
@@ -1,3 +1,5 @@
+package org.altumtek;
+
 import java.io.IOException;
 import java.net.*;
