+++ content-searching/src/main/java/org/altumtek/Node.java
@@ -1,3 +1,5 @@
+package org.altumtek;
+
 /**
  * Created by chanuka on 10/12/18.
  */
